Add an example showing how to install git and ssh
As the default executor used by the shellcheck orb is lacking git
and ssh, CircleCI will fall back to its own implementation.

While this works fine for many cases, there are a few for which
this does not work. One of those is execution of CI on git tags.

A workaround which works well for us is to pre-install the
[required tools](https://circleci.com/docs/2.0/custom-images/#required-tools-for-primary-containers)
in those cases.